Split out sidebar components into separate files
